# derp: log highest bandwidth ip