@JulianKnodt 2021-06-16
Benchstat:
```
name                      old time/op    new time/op    delta
SendRecv/msgsize=10-8       8.95µs ± 5%    8.22µs ±20%   ~     (p=0.421 n=5+5)
SendRecv/msgsize=100-8      9.10µs ± 1%    9.37µs ±21%   ~     (p=1.000 n=4+5)
SendRecv/msgsize=1000-8     10.4µs ± 6%    10.1µs ±19%   ~     (p=0.310 n=5+5)
SendRecv/msgsize=10000-8    28.5µs ± 3%    27.0µs ±14%   ~     (p=0.151 n=5+5)

name                      old speed      new speed      delta
SendRecv/msgsize=10-8     1.12MB/s ± 4%  1.23MB/s ±23%   ~     (p=0.341 n=5+5)
SendRecv/msgsize=100-8    10.6MB/s ±13%  10.9MB/s ±24%   ~     (p=0.841 n=5+5)
SendRecv/msgsize=1000-8   96.4MB/s ± 6%  99.5MB/s ±17%   ~     (p=0.310 n=5+5)
SendRecv/msgsize=10000-8   351MB/s ± 3%   373MB/s ±15%   ~     (p=0.151 n=5+5)

name                      old alloc/op   new alloc/op   delta
SendRecv/msgsize=10-8         112B ± 0%      112B ± 0%   ~     (all equal)
SendRecv/msgsize=100-8        208B ± 0%      207B ± 1%   ~     (p=0.333 n=4+5)
SendRecv/msgsize=1000-8     1.12kB ± 0%    1.12kB ± 0%   ~     (p=0.333 n=4+5)
SendRecv/msgsize=10000-8    19.9kB ± 1%    19.8kB ± 1%   ~     (p=0.690 n=5+5)

name                      old allocs/op  new allocs/op  delta
SendRecv/msgsize=10-8         3.00 ± 0%      2.60 ±23%   ~     (p=0.333 n=4+5)
SendRecv/msgsize=100-8        3.00 ± 0%      2.60 ±23%   ~     (p=0.333 n=4+5)
SendRecv/msgsize=1000-8       3.00 ± 0%      2.60 ±23%   ~     (p=0.444 n=5+5)
SendRecv/msgsize=10000-8      3.00 ± 0%      3.00 ± 0%   ~     (all equal)
```