@JulianKnodt 2021-06-18
Closed in favor of a different implementation which maps IP --> public Key for logging with unix tools